requirements.txt
PySide6
PyYAML>=6.0
requests>=2.28.0
beautifulsoup4>=4.11.1

